requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
Pillow>=9.3.0
geopy==2.2.0
